# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kombucha Tracker API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      security:
        - clerkAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile / complete onboarding
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
...
  /batches/{batchId}/logs:
    get:
      operationId: listLogs
      tags: [logs]
      summary: List daily logs for a batch
      security:
        - clerkAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Log"
        "401":
          description: Unauthorized
        "404":
          description: Batch not found
    post:
      operationId: createLog
      tags: [logs]
      summary: Create a daily log entry for a batch
      security:
        - clerkAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLogBody"
      responses:
        "201":
          description: Created log with AI tip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Log"
        "401":
          description: Unauthorized
  /batches/{batchId}/logs/{logId}:
    get:
      operationId: getLog
      tags: [logs]
      summary: Get a single log entry
      security:
        - clerkAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
        - name: logId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Log entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Log"
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /batches/{batchId}/photos:
    get:
      operationId: listPhotos
      tags: [photos]
      summary: List photos for a batch
      security:
        - clerkAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
        "401":
          description: Unauthorized
    post:
      operationId: createPhoto
      tags: [photos]
      summary: Register a photo after upload (save metadata)
      security:
        - clerkAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePhotoBody"
      responses:
        "201":
          description: Created photo record with AI analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "401":
          description: Unauthorized
  /batches/{batchId}/photos/{photoId}:
    delete:
      operationId: deletePhoto
      tags: [photos]
      summary: Delete a photo
      security:
        - clerkAuth: []
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        hasCompletedOnboarding:
          type: boolean
        hasMadeBefore:
          type: boolean
        hasScoby:
          type: boolean
        currentStage:
          type: string
          nullable: true
        experienceLevel:
          type: string
          nullable: true
        flavorPreference:
          type: string
          nullable: true
        onboardingAdvice:
          type: string
          nullable: true
        ttsEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkUserId
        - hasCompletedOnboarding
        - hasMadeBefore
        - hasScoby
        - ttsEnabled
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        hasCompletedOnboarding:
          type: boolean
        hasMadeBefore:
          type: boolean
        hasScoby:
          type: boolean
        currentStage:
          type: string
        experienceLevel:
          type: string
        flavorPreference:
          type: string
        onboardingAdvice:
          type: string
        ttsEnabled:
          type: boolean
...
    Log:
      type: object
      properties:
        id:
          type: integer
        batchId:
          type: integer
        dayNumber:
          type: integer
        temperature:
          type: number
          nullable: true
        scobylook:
          type: string
          nullable: true
        smell:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        aiTip:
          type: string
          nullable: true
        loggedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - batchId
        - dayNumber
        - loggedAt
        - createdAt
    CreateLogBody:
      type: object
      properties:
        dayNumber:
          type: integer
        temperature:
          type: number
        scobylook:
          type: string
        smell:
          type: string
          enum: [good, sour, vinegary, strange]
        color:
          type: string
        notes:
          type: string
        loggedAt:
          type: string
          format: date-time
      required:
        - dayNumber
    Photo:
      type: object
      properties:
        id:
          type: integer
        batchId:
          type: integer
        objectPath:
          type: string
        caption:
          type: string
          nullable: true
        dayNumber:
          type: integer
          nullable: true
        aiAnalysis:
          type: string
          nullable: true
        takenAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - batchId
        - objectPath
        - takenAt
        - createdAt
    CreatePhotoBody:
      type: object
      properties:
        objectPath:
          type: string
        caption:
          type: string
        dayNumber:
          type: integer
        takenAt:
          type: string
          format: date-time
        analyzeWithAi:
          type: boolean
      required:
        - objectPath
...