# lib/api-zod/src/generated/types/createLogBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kombucha Tracker API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateLogBodySmell } from "./createLogBodySmell";

export interface CreateLogBody {
  dayNumber: number;
  temperature?: number;
  scobylook?: string;
  smell?: CreateLogBodySmell;
  color?: string;
  notes?: string;
  loggedAt?: Date;
}