# lib/api-zod/src/generated/types/log.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kombucha Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Log {
  id: number;
  batchId: number;
  dayNumber: number;
  temperature?: number | null;
  scobylook?: string | null;
  smell?: string | null;
  color?: string | null;
  notes?: string | null;
  aiTip?: string | null;
  loggedAt: Date;
  createdAt: Date;
}