# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kombucha Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  clerkUserId: string;
  hasCompletedOnboarding: boolean;
  hasMadeBefore: boolean;
  hasScoby: boolean;
  currentStage?: string | null;
  experienceLevel?: string | null;
  flavorPreference?: string | null;
  onboardingAdvice?: string | null;
  ttsEnabled: boolean;
  createdAt: Date;
}